+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Unreleased changes (dev)
+- Added ability to swap team colors in a match. #49 - NicEastvillage
+
 #### Version 1.2.0 - 31. Jan 2019
 - Double elimination. #9 - NicEastvillage
     - Supports any number of teams and byes